requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
